experimenting with new directory travel

+++ fat32.h
@@ -27,11 +27,11 @@ typedef struct fat32_struct
 /*** public functions ***/
 fat32* createFat32(int deviceFP);
 
-void readCluster(fat32 *fat32Obj, int clusterNum, char buf[]);
+void readCluster(fat32 *fat32Obj, uint64_t clusterNum, char buf[]);
 
-uint32_t readFAT(fat32 *fat32Obj, int n);
+uint32_t readFAT(fat32 *fat32Obj, uint32_t n);
 
-void readSector(fat32 *fat32Obj, int sectorNum, char buf[]);
+void readSector(fat32 *fat32Obj, uint64_t sectorNum, char buf[]);
 
 void checkFat32(fat32 *fat32Obj);
